+++ app/admin/course_events.rb
@@ -129,8 +129,8 @@ ActiveAdmin.register CourseEvent do
           ff.semantic_errors *ff.object.errors.keys
           tmp_params = current_admin_user.is_root? ? nil : { 'q[organization_id_equals]' => f.object.organization_id }
           ff.input :user_id, as: :search_select, url: admin_users_path(tmp_params),
-                  fields: [:email], display_name: 'email', minimum_input_length: 3,
-                  order_by: 'email_asc'
+                   fields: [:firstname, :lastname], display_name: :full_name, minimum_input_length: 3,
+                  order_by: 'lastname_asc'
           ff.input :presence
         end
       end