+++ app/admin/users.rb
@@ -164,8 +164,8 @@ ActiveAdmin.register User do
       f.input :confirmed_at, as: :datetime_picker
       tmp_params = current_admin_user.is_root? ? nil : { 'q[organization_id_equals]' => f.object.organization_id }
       f.input :trainer_id, as: :search_select, url: admin_trainers_path(tmp_params),
-              fields: [:nickname], display_name: 'nickname', minimum_input_length: 3,
-              order_by: 'nickname_asc'
+              fields: [:lastname], display_name: 'lastname', minimum_input_length: 3,
+              order_by: 'lastname_asc'
     end
 
     f.actions