+++ app/controllers/api/ui/v1/user_documents_controller.rb
@@ -3,6 +3,8 @@ class Api::Ui::V1::UserDocumentsController < Api::Ui::BaseController
 
   respond_to :json
 
+  # PUT from external service to update the state and the sign_checksum
+  # Example of a sign checksum: 9d1f97c55893d83fa241ef73fd2612e7
   def callback
     user_document = UserDocument.find_by_uuid(user_document_params[:uuid])
     raise(ActiveRecord::RecordNotFound, "Uuid #{user_document_params[:uuid]} does not exist" ) unless user_document
@@ -15,7 +17,8 @@ class Api::Ui::V1::UserDocumentsController < Api::Ui::BaseController
       user_document.errors.add :state, "Unexpected state #{params[:state]}"
     end
 
-    if user_document.errors.empty? && user_document.update(state: user_document_params[:state])
+    if user_document.errors.empty? && user_document.update(state: user_document_params[:state],
+                                                           sign_checksum: user_document_params[:final_checksum])
       render json: { state: user_document.state }, status: :ok
     else
       render json: { errors: user_document.errors }, status: :unprocessable_entity
@@ -25,6 +28,6 @@ class Api::Ui::V1::UserDocumentsController < Api::Ui::BaseController
   private
 
   def user_document_params
-    params.permit(:uuid, :state, :format, :user_document)
+    params.permit(:uuid, :state, :format, :user_document, :final_checksum)
   end
 end