+++ db/migrate/20200910211720_add_sign_checksum_to_user_document.rb
@@ -1,0 +1,7 @@
+class AddSignChecksumToUserDocument < ActiveRecord::Migration[6.0]
+  def change
+    add_column :user_documents, :sign_checksum, :string, limit: 32
+    add_index :users, :firstname
+    add_index :users, :lastname
+  end
+end